requirements.txt
requests==2.25.1
urllib3==1.26.3
beautifulsoup4==4.9.3
pycryptodome==3.10.1
